Implement removeTableByKey
