+++ lib/changes/removeTable.js
@@ -1,8 +1,8 @@
 // @flow
-import { Block, Text, type Change } from 'slate';
+import { type Change } from 'slate';
 
-import { TablePosition } from '../utils';
 import type Options from '../options';
+import removeTableByKey from './removeTableByKey';
 
 /**
  * Delete the whole table at position
@@ -11,45 +11,7 @@ function removeTable(opts: Options, change: Change): Change {
     const { value } = change;
     const { startKey } = value;
 
-    const pos = TablePosition.create(opts, value.document, startKey);
-    const { table } = pos;
-    const { document } = change.value;
-    let nextFocusBlock = null;
-    let shouldCollapseToEnd = false;
-
-    const nextBlock = change.value.document.getNextBlock(table.key);
-    if (nextBlock) {
-        nextFocusBlock = nextBlock;
-    } else {
-        const prevBlock = change.value.document.getPreviousBlock(table.key);
-        if (prevBlock) {
-            nextFocusBlock = prevBlock;
-            shouldCollapseToEnd = true;
-        } else if (opts.exitBlockType) {
-            nextFocusBlock = Block.create({
-                type: opts.exitBlockType,
-                nodes: [Text.create('')]
-            });
-            const tableParent = document.getParent(table.key);
-            const insertionIndex = tableParent.nodes.indexOf(table) + 1;
-            change.insertNodeByKey(
-                tableParent.key,
-                insertionIndex,
-                nextFocusBlock
-            );
-        }
-    }
-
-    change.removeNodeByKey(table.key);
-    if (!nextFocusBlock) {
-        return change;
-    }
-    if (shouldCollapseToEnd) {
-        change.collapseToEndOf(nextFocusBlock).focus();
-    } else {
-        change.collapseToStartOf(nextFocusBlock).focus();
-    }
-    return change;
+    return removeTableByKey(opts, change, startKey);
 }
 
 export default removeTable;
